refactor(tests): Consolidate permission tests and add coverage tooling
- Merge permissions-roster.spec.ts into permissions.spec.ts
  - Removed 10 duplicate tests for registered member permissions
  - Kept 3 unique tests for registration checks (email_id=null)
- Add vitest coverage configuration with @vitest/coverage-v8
- Add test:coverage script for on-demand coverage reports

Before: 59 test files, 785 tests, 16,846 lines
After:  58 test files, 778 tests, 16,730 lines

Reduction: 1 file, 7 tests, 116 lines of test code

Coverage: 87.77% statements, 87.21% branches

Refs: #146
