+++ apps/vault/package.json
@@ -10,6 +10,7 @@
 		"prepare": "svelte-kit sync || echo ''",
 		"test": "vitest run",
 		"test:watch": "vitest",
+		"test:coverage": "vitest run --coverage",
 		"test:e2e": "playwright test",
 		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
 		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
@@ -25,6 +26,7 @@
 		"@tailwindcss/vite": "^4.1.2",
 		"@testing-library/svelte": "^5.2.4",
 		"@types/node": "^22.10.2",
+		"@vitest/coverage-v8": "^3.2.4",
 		"eslint": "^9.17.0",
 		"eslint-config-prettier": "^9.1.0",
 		"eslint-plugin-svelte": "^2.46.1",
